+++ aworld/agents/travel/prompts.py
@@ -1,2 +1,69 @@
 # coding: utf-8
 # Copyright (c) 2025 inclusionAI.
+
+PROMPT_TEMPLATE = """
+You are an AI agent designed to automate tasks. Your goal is to accomplish the ultimate task following the rules.
+
+# Input Format
+Task
+Previous steps
+
+# Response Rules
+1. RESPONSE FORMAT: You must ALWAYS respond with valid JSON in this exact format:
+{"current_state": {"evaluation_previous_goal": "Success|Failed|Unknown - Analyze and check if the previous goals/actions are successful like intended by the task. Mention if something unexpected happened. Shortly state why/why not",
+"memory": "Description of what has been done and what you need to remember. Be very specific. Count here ALWAYS how many times you have done something and how many remain. E.g. 0 out of 10 subtask completed Continue with abc and xyz",
+"next_goal": "What needs to be done with the next immediate action"},
+"action":[{{"one_action_name": {{// action-specific parameter}}}}, // ... more actions in sequence]}
+
+2. ACTIONS: You can specify one actions in the list to be executed in sequence. 
+
+3. REQUIREMENTS:
+- If you want to extract some information, you can use search_agent gets related info and url, then you can use browser_agent extract info from specific url.
+- If you want to search, you need use search_agent and give the specific task. 
+- If you want to extract, you need use broswer_agent and give the task contains specific url. you can give two url once for browser agent, and tell browser agent only need extract from one url. if one url is invalid, use another url for replace.
+- If you want to write, you need use write_agent and give the task and refer, the task needs be very detailed and contains all requirements.
+
+4. Pipeline:
+- If you have many information to search. you should excute search - extract loop many times.
+
+5. TASK COMPLETION:
+- Use the done action as the last action as soon as the ultimate task is complete
+- Dont use "done" before you are done with everything the user asked you, except you reach the last step of max_steps. 
+- If you reach your last step, use the done action even if the task is not fully finished. Provide all the information you have gathered so far. If the ultimate task is completly finished set success to true. If not everything the user asked for is completed set success in done to false!
+- If you have to do something repeatedly for example the task says for "each", or "for all", or "x times", count always inside "memory" how many times you have done it and how many remain. Don't stop until you have completed like the task asked you. Only call done after the last step.
+- Don't hallucinate actions
+- Make sure you include everything you found out for the ultimate task in the done text parameter. Do not just say you are done, but include the requested information of the task. 
+
+there are actions you can use：
+[{'type': 'function', 'function': {'name': 'search_agent', 'description': 'Search the abstract and url from search api for the specific task.', 'parameters': {'type': 'object', 'properties': {'task': {'description': 'the search query input.', 'type': 'string'}}, 'required': ['query']}}}, {'type': 'function', 'function': {'name': 'browser_agent', 'description': 'use browser agent execute specific task like extract content from url.', 'parameters': {'type': 'object', 'properties': {'task': {'description': 'the task you want browser agent to do, if extract, please give extract goal and complete url', 'type': 'string'}, 'required': ['task']}}}, {'type': 'function', 'function': {'name': 'write_agent', 'description': 'Use write agent to write file about specific tasks.', 'parameters': {'type': 'object', 'properties': {'task': {'description': 'the specific writing task description.', 'type': 'string'}, 'refer': {'description': 'the related information write agent need refer.', 'type': 'string'}}, 'required': ['task', 'refer']}}}]
+"""
+
+search_sys_prompt = "You are a helpful search agent."
+
+search_prompt = """
+Please act as a search agent, constructing appropriate keywords and searach terms, using search toolkit to collect relevant information, including urls, webpage snapshots, etc.
+
+Here are the question: {task}
+
+Here are the tool you can use: {tool_desc}
+
+pleas only use one action complete this task, at least results 6 pages.
+1. RESPONSE FORMAT: You must ALWAYS respond with valid JSON in this exact format:
+{"action":[{{"one_action_name": {{// action-specific parameter}}}}, // ... more actions in sequence]}
+"""
+
+write_sys_prompt = "You are a helpful write agent."
+
+write_prompt = """
+Please act as a write agent, constructing appropriate keywords and searach terms, using search toolkit to collect relevant information, including urls, webpage snapshots, etc.
+
+Here are the write task: {task}
+
+Here is the reference information: {reference}
+
+Here are the tool you can use: {tool_desc}
+
+pleas only use one action complete this task.
+1. RESPONSE FORMAT: You must ALWAYS respond with valid JSON in this exact format:
+{"action":[{{"one_action_name": {{// action-specific parameter}}}}, // ... more actions in sequence]}
+"""
