+++ aworld/agents/travel/search_agent.py
@@ -3,6 +3,7 @@ import os
 import json
 from typing import Dict, Any, List, Union
 
+from aworld.agents.travel.prompts import search_prompt, search_sys_prompt
 from aworld.config.common import Agents, Tools
 from aworld.config.conf import AgentConfig
 from aworld.core.agent.base import AgentFactory, BaseAgent
@@ -11,23 +12,6 @@ from aworld.core.envs.tool_desc import get_tool_desc_by_name
 from aworld.models.utils import tool_desc_transform
 from aworld.core.envs.tool import ToolFactory
 from aworld.config.conf import load_config
-
-
-sys_prompt = "You are a helpful search agent."
-
-prompt = """
-Please act as a search agent, constructing appropriate keywords and searach terms, using search toolkit to collect relevant information, including urls, webpage snapshots, etc.
-
-Here are the question: {task}
-
-Here are the tool you can use: {tool_desc}
-
-pleas only use one action complete this task, at least results 6 pages.
-"""
-
-response_format = """1. RESPONSE FORMAT: You must ALWAYS respond with valid JSON in this exact format:
-{"action":[{{"one_action_name": {{// action-specific parameter}}}}, // ... more actions in sequence]}
-"""
 
 
 # Step1
@@ -49,9 +33,9 @@ class SearchAgent(BaseAgent):
 
             return [ActionModel(tool_name="[done]", policy_info=observation.content)]
 
-        messages = [{'role': 'system', 'content': sys_prompt},
+        messages = [{'role': 'system', 'content': search_sys_prompt},
                     {'role': 'user',
-                     'content': prompt.format(task=observation.content, tool_desc=self.tool_desc) + response_format}]
+                     'content': search_prompt.format(task=observation.content, tool_desc=self.tool_desc)}]
 
         llm_result = self.llm.invoke(
             input=messages,
@@ -67,16 +51,11 @@ class SearchAgent(BaseAgent):
         parsed_results = []
 
         for action in actions:
-            # 遍历 action 中的键值对
             for key, value in action.items():
-                # 分割 action_name 和 tool_name
                 if "__" in key:
                     tool_name, action_name = key.split("__", 1)
 
-                # 提取 params
                 params = value
-
-                # 将解析结果存入列表
                 parsed_results.append(ActionModel(tool_name=tool_name,
                                                   action_name=action_name,
                                                   params=params))
