+++ aworld/agents/travel/write_agent.py
@@ -3,6 +3,7 @@ import os
 import json
 from typing import Dict, Any, List, Union
 
+from aworld.agents.travel.prompts import write_prompt, write_sys_prompt
 from aworld.config.conf import AgentConfig
 from aworld.core.agent.base import AgentFactory, BaseAgent
 from aworld.core.common import Observation, ActionModel
@@ -10,24 +11,6 @@ from aworld.core.envs.tool_desc import get_tool_desc_by_name
 from aworld.models.utils import tool_desc_transform
 from aworld.core.envs.tool import ToolFactory
 from aworld.config.conf import load_config
-
-sys_prompt = "You are a helpful write agent."
-
-prompt = """
-Please act as a write agent, constructing appropriate keywords and searach terms, using search toolkit to collect relevant information, including urls, webpage snapshots, etc.
-
-Here are the write task: {task}
-
-Here is the reference information: {reference}
-
-Here are the tool you can use: {tool_desc}
-
-pleas only use one action complete this task.
-"""
-
-response_format = """1. RESPONSE FORMAT: You must ALWAYS respond with valid JSON in this exact format:
-{"action":[{{"one_action_name": {{// action-specific parameter}}}}, // ... more actions in sequence]}
-"""
 
 
 @AgentFactory.register(name='write_agent', desc="write agent")
@@ -49,8 +32,8 @@ class WriteAgent(BaseAgent):
 
             return [ActionModel(tool_name="[done]", policy_info=observation.content)]
 
-        messages = [{'role': 'system', 'content': sys_prompt},
-                    {'role': 'user', 'content': prompt.format(task=observation.content['task'], reference=observation.content['refer'], tool_desc=self.tool_desc) + response_format}]
+        messages = [{'role': 'system', 'content': write_sys_prompt},
+                    {'role': 'user', 'content': write_prompt.format(task=observation.content['task'], reference=observation.content['refer'], tool_desc=self.tool_desc) + response_format}]
 
         llm_result = self.llm.invoke(
             input=messages,
@@ -72,16 +55,11 @@ class WriteAgent(BaseAgent):
         parsed_results = []
 
         for action in actions:
-            # 遍历 action 中的键值对
             for key, value in action.items():
-                # 分割 action_name 和 tool_name
                 if "__" in key:
                     tool_name, action_name = key.split("__", 1)
 
-                # 提取 params
                 params = value
-
-                # 将解析结果存入列表
                 parsed_results.append(ActionModel(tool_name=tool_name,
                                        action_name=action_name,
                                        params=params))
